requirements.txt
# Core dependencies for Sports Calendar Manager MVP
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
click>=8.0.0
python-dotenv>=1.0.0
pytz>=2023.3
streamlit>=1.28.0
pyyaml>=6.0.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
